Improve integer precision to uint64 as ovs does. We want to generate interface metrics from interface statictis and found that precision in libovsdb and ovs not match. https://sourcegraph.com/github.com/openvswitch/ovs@d29491eeb4a6e2128430eeac9a08327d8e7d8ed6/-/blob/lib/netdev-linux.c?L204:5

+++ cache/cache_test.go
@@ -18,7 +18,7 @@ type testModel struct {
 	UUID  string   `ovsdb:"_uuid"`
 	Foo   string   `ovsdb:"foo"`
 	Bar   string   `ovsdb:"bar"`
-	Baz   int      `ovsdb:"baz"`
+	Baz   uint64   `ovsdb:"baz"`
 	Array []string `ovsdb:"array"`
 }
